Tests over isResponseErrorWithStatus, and praise the creator I found where Shoup buried code that creates a valid ResponseError.
